emailvalidator/emailvalidator/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "emailvalidator"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "email-validator>=2.1.0",      # Layer 1 — syntax
    "dnspython>=2.6.0",            # Layer 2 — DNS/MX
    "redis>=5.0.0",                # queue + caching
    "celery>=5.4.0",               # worker queue
    "sqlalchemy>=2.0.0",           # result persistence
    "asyncpg>=0.29.0",             # async postgres driver
    "alembic>=1.13.0",             # migrations
    "pydantic-settings>=2.0.0",    # config/env
    "httpx>=0.27.0",               # async HTTP client
    "tenacity>=8.3.0",             # retry logic for SMTP
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
